requirements.txt
numpy
biopython
pandas
tqdm
psutil
numba
pysam